Problem: atomic_counter_t can't be inlined in union
Solution: refactor class to use common struct + functions.
This is necessary in order to inline content_t in the lmsg type, and
get rid of one malloc per message.
The existing class is left in place and just calls the new functions,
to avoid breaking internal API and generating extensive changes.