Problem: zero-copy API does one malloc per message
Solution: inline the content of content_t struct inside the lsmg
union member to get rid of the extra malloc per message.
Even if to do this it was necessary to bump the message size from
64 to 128 bytes, thus occupying 2 cache lines instead of 1 which is
better for performances, getting rid of one extra malloc per message
more than makes up for it.