Problem: creating monitor without connecting it hangs socket
Solution: set send timeout of 0 in inproc monitor socket.
Socket events are already lossy: if a socket monitor is not set up,
no events will be generated. Extend it to drop messages if there
is no peer as well to avoid deadlocking sockets. Add a test case
to verify the fix.