Problem: race condition at shutdown in random.cpp
Solution: avoid using the zmq::mutex class, which has a non-trivial
destructor, as a global static variable in random.ccp. This causes a
race between the context and the mutex destructors as the order is not
defined.
Instead use a simple pthread mutex on *NIX, which can be statically
initialised and does not need to be de-initialised, removing the race.
TODO: a similar solution needs to be implemented for Windows.

Fixes #2915